Moved remaining AnimatableValue parsing to Parser classes
