+++ lottie/src/main/java/com/airbnb/lottie/parser/CircleShapeParser.java
@@ -4,7 +4,6 @@ import android.graphics.PointF;
 import android.util.JsonReader;
 
 import com.airbnb.lottie.LottieComposition;
-import com.airbnb.lottie.model.animatable.AnimatablePathValue;
 import com.airbnb.lottie.model.animatable.AnimatablePointValue;
 import com.airbnb.lottie.model.animatable.AnimatableValue;
 import com.airbnb.lottie.model.content.CircleShape;
@@ -28,8 +27,7 @@ public class CircleShapeParser {
           name = reader.nextString();
           break;
         case "p":
-          position = AnimatablePathValue
-              .createAnimatablePathOrSplitDimensionPath(reader, composition);
+          position = AnimatablePathValueParser.parseSplitPath(reader, composition);
           break;
         case "s":
           size = AnimatableValueParser.parsePoint(reader, composition);
