+++ lottie/src/main/java/com/airbnb/lottie/parser/DocumentDataParser.java
@@ -3,7 +3,6 @@ package com.airbnb.lottie.parser;
 import android.util.JsonReader;
 
 import com.airbnb.lottie.model.DocumentData;
-import com.airbnb.lottie.utils.JsonUtils;
 
 import java.io.IOException;
