+++ lottie/src/main/java/com/airbnb/lottie/parser/FloatParser.java
@@ -2,8 +2,6 @@ package com.airbnb.lottie.parser;
 
 import android.util.JsonReader;
 
-import com.airbnb.lottie.utils.JsonUtils;
-
 import java.io.IOException;
 
 public class FloatParser implements ValueParser<Float> {
